+++ src/main/matlab/+edu/+washington/+riekelab/+patterson/+protocols/GratingDirection.m
@@ -1,5 +1,4 @@
 classdef GratingDirection < edu.washington.riekelab.protocols.RiekeLabStageProtocol
-% 2020-09-13    SSP
 
     properties 
         amp
@@ -13,6 +12,7 @@ classdef GratingDirection < edu.washington.riekelab.protocols.RiekeLabStageProto
         directions = 0:30:330           % Grating direction (degrees) 
         spatialFrequency = 2            % Cycles per short axis of screen
         backgroundIntensity = 0.5       % Mean light level (0-1)
+        apertureDiameter = 0            % Diameter of aperture (pixels)
         onlineAnalysis = 'none'         % Analysis type
         randomOrder = false             % Randomize epochs?
         numberOfAverages = uint16(36)   % Number of epochs         
@@ -38,16 +38,6 @@ classdef GratingDirection < edu.washington.riekelab.protocols.RiekeLabStageProto
             [obj.amp, obj.ampType] = obj.createDeviceNamesProperty('Amp');
         end
 
-        function p = getPreview(obj, panel)
-            if isempty(obj.rig.getDevices('Stage'))
-                p = [];
-                return
-            end
-            p = io.github.stage_vss.previews.StagePreview(panel,...
-                @()obj.createPresentation(),...
-                'windowSize', obj.rig.getDevice('Stage').getCanvasSize());
-        end
-
         function prepareRun(obj)
             prepareRun@edu.washington.riekelab.protocols.RiekeLabStageProtocol(obj);
 
@@ -64,7 +54,8 @@ classdef GratingDirection < edu.washington.riekelab.protocols.RiekeLabStageProto
 
             % Set up figures
             rgb = edu.washington.riekelab.patterson.utils.multigradient(...
-                'preset', 'div.cb.spectral.9', 'length', numel(obj.directions));
+                'preset', 'div.cb.spectral.9',... 
+                'length', numel(unique(obj.directions)));
 
             obj.showFigure('symphonyui.builtin.figures.ResponseFigure',...
                 obj.rig.getDevice(obj.amp));
@@ -73,18 +64,21 @@ classdef GratingDirection < edu.washington.riekelab.protocols.RiekeLabStageProto
                 'recordingType', obj.onlineAnalysis, 'sweepColor', rgb);
             obj.showFigure('edu.washington.riekelab.patterson.figures.FrameTimingFigure',...
                 obj.rig.getDevice('Stage'), obj.rig.getDevice('Frame Monitor'));
+            
             if ~strcmp(obj.onlineAnalysis, 'none')
                 obj.showFigure('edu.washington.riekelab.patterson.figures.F1F2Figure',...
-                    obj.rig.getDevice(obj.amp), obj.directions, obj.onlineAnalysis,...
-                    obj.preTime, obj.stimTime, 'waitTime', obj.waitTime,...
-                    'temporalFrequency', obj.temporalFrequency,...
-                    'xName', 'direction', 'showF2', false);
+                    obj.rig.getDevice(obj.amp), obj.onlineAnalysis,...
+                    obj.preTime, obj.stimTime, 'WaitTime', obj.waitTime,...
+                    'TemporalFrequency', obj.temporalFrequency,...
+                    'VariedParameterName', 'direction',...
+                    'GraphName', 'Direction Tuning');
             end
         end
 
         function p = createPresentation(obj)
             device = obj.rig.getDevice('Stage');
             canvasSize = device.getCanvasSize();
+            centerOffsetPix = obj.rig.getDevice('Stage').um2pix(obj.centerOffset);
 
             p = stage.core.Presentation((obj.preTime + obj.waitTime + obj.stimTime + obj.tailTime) * 1e-3);
             p.setBackgroundColor(obj.backgroundIntensity);
@@ -104,6 +98,17 @@ classdef GratingDirection < edu.washington.riekelab.protocols.RiekeLabStageProto
             imgController = stage.builtin.controllers.PropertyController(grate, 'imageMatrix',...
                 @(state)getGratingDrift(obj, state.time - (obj.preTime + obj.waitTime) * 1e-3));
             p.addController(imgController);
+
+            if obj.apertureDiameter > 0
+                aperture = stage.builtin.stimuli.Rectangle();
+                aperture.position = canvasSize/2 + centerOffsetPix;
+                aperture.color = obj.backgroundIntensity;
+                aperture.size = 2 * max(canvasSize) * ones(1, 2);
+                mask = stage.core.Mask.createCircularAperture(...
+                    obj.apertureDiameter / (2 * max(canvasSize)));
+                aperture.setMask(mask);
+                p.addStimulus(aperture);
+            end
 
             function g = getGratingDrift(obj, time)
                 if time >= 0