+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # CHANGELOG
 
 ### 20200916
-- Added `TestModulation` that runs on `SimulatedConfocal` and demos `F1F2Figure` with real data. To Do: remove dependency on `SineGenerator` for simplicity
+- Added `TestModulation` that runs on `SimulatedConfocal` and demos `F1F2Figure` with real data. 
+- Removed custom `SineGenerator` for simplicity
+- Improved `F1F2Figure`
+- Added aperture option to grating stimuli
 
 ### 20200915
 - Added `BarDirection` and `BarSpeed`